requirements.txt
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
scikit-learn==1.3.2
xgboost==2.0.0
lightgbm==4.3.0
shap==0.43.0
scipy==1.11.1
plotly==5.15.0
streamlit==1.28.0
jupyter==1.0.0
ipykernel==6.25.0